requirements.txt
python-dotenv==1.0.0
streamlit==1.28.1
plotly==5.18.0
pandas>=2.1.3
